+++ packages/nodes/src/components/nodes/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useDraggableNode } from './useDraggableNode'
+export { usePort } from './usePort'
